sort categories as mixin
